SotaWhisperTest/requirements.txt
torch
torchaudio
openai-whisper
flask
pyttsx3
